IMP: dynamic load of setup functions
